+++ NEWS.md
@@ -1,3 +1,5 @@
+# utilscognigen (development version)
+
 # utilscognigen 1.0.2
 
 * Updated to host publicly on GitHub.