+++ NEWS.md
@@ -1,4 +1,4 @@
-# utilscognigen (development version)
+# utilscognigen 1.0.3
 
 * Now accepts Quarto documents as input where R or R Markdown files are accepted.
 * Added functions to view project management sites: `browse_project_gantt` and `browse_project_tracker`.